pyproject.toml
[project]
name = "anita"
version = "0.1.0"
description = ""
authors = [
    {name = "timpara",email = "t.paraskevopoulos@posteo.de"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "genanki (>=0.13.1,<0.14.0)",
    "openai (>=1.95.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "elevenlabs (>=2.7.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
